Update: Rearrange Map for prop passing purposes.
