+++ src/Components/App/App.js
@@ -1,10 +1,10 @@
 import './App.css';
-import Map from '../Map/Map';
 import Form from '../Form/Form';
 import Review from '../Review/Review';
 import { useQuery, gql } from '@apollo/client';
+import { useState } from 'react';
+import { MapContainer, TileLayer, Marker, Popup, useMapEvents } from 'react-leaflet';
 
-const App = () => {
 const GET_REVIEWS = gql`
   query {
     reviews {
@@ -16,15 +16,49 @@ const GET_REVIEWS = gql`
     }
   }
 `;
+
+const App = () => {
 const { loading, error, data } = useQuery(GET_REVIEWS);
+const [position, setPosition] = useState(null)
+const [lat, setLat] = useState(null)
+const [lng, setLng] = useState(null)
+
+const LocationMarker = () => {
+  const map = useMapEvents({
+    click(e) {
+      const { lat, lng } = e.latlng;
+      setPosition({ lat, lng });
+      setLat(lat.toString())
+      setLng(lng.toString())
+      console.log(typeof lat)
+      map.flyTo(e.latlng, map.getZoom());
+    }
+  })
+
+  return position === null ? null : (
+    <Marker position={position}>
+      <Popup>Selected Position</Popup>
+    </Marker>
+  );
+};
+
 if (loading) return <p>Loading...</p>;
 if (error) return <p>Error : {error.message}</p>;
 console.log('this is data', data)
 
   return (
     <div className='app'>
-      <Form />
-      <Map />
+      <Form lat={lat} lng={lng}/>
+      <MapContainer
+      center={{ lat: 39.739235, lng: -104.990250 }}
+      zoom={6}
+      >
+        <TileLayer
+          attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
+          url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+        />
+        <LocationMarker />
+      </MapContainer>
       <Review data={data}/>
     </div>
   );
