+++ src/Components/Review/Review.js
@@ -2,6 +2,7 @@ import React from 'react';
 
 const Review = ({ data }) => {
 
+  
   return (
     <div className='review-container'>
       {data.reviews.length ? (
@@ -10,6 +11,8 @@ const Review = ({ data }) => {
             <div className='review-card' key={review.id}>
               <div>Name: {review.name}</div>
               <div>Description: {review.description}</div>
+              <div>Latitude: {review.lat}</div>
+              <div>Longitude: {review.lng}</div>
             </div>
           ))}
         </div>
